pyproject.toml
[tool.poetry]
name = "googleadsquerytool"
version = "0.1.4"
description = "This is a package you can use to query reporting data from the Google Ads API."
authors = ["caspercrause <ccrause07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
google-ads = "==22.1.0"
pandas = ">1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
